+++ packages/widget/src/components/Pages/Swap/Form/FeeDetails/index.tsx
@@ -45,15 +45,17 @@ export default function QuoteDetails({ swapValues: values, quote, isQuoteLoading
                 quote &&
                 <Accordion type='single' collapsible className='w-full' value={isAccordionOpen ? 'quote' : ''} onValueChange={(value) => { setIsAccordionOpen(value === 'quote') }}>
                     <AccordionItem value='quote' className='bg-secondary-500 rounded-2xl'>
-                        <AccordionTrigger className={clsx(
-                            'p-3.5 pr-5 w-full rounded-2xl flex items-center justify-between transition-colors duration-200 hover:bg-secondary-400 mt-3',
-                            triggerClassnames,
-                            {
-                                'bg-secondary-500': !isAccordionOpen,
-                                'bg-secondary-400': isAccordionOpen,
-                                'animate-pulse-strong': isQuoteLoading && !isAccordionOpen
-                            }
-                        )}>
+                        <AccordionTrigger
+                            data-attr="see-swap-details"
+                            className={clsx(
+                                'p-3.5 pr-5 w-full rounded-2xl flex items-center justify-between transition-colors duration-200 hover:bg-secondary-400 mt-3',
+                                triggerClassnames,
+                                {
+                                    'bg-secondary-500': !isAccordionOpen,
+                                    'bg-secondary-400': isAccordionOpen,
+                                    'animate-pulse-strong': isQuoteLoading && !isAccordionOpen
+                                }
+                            )}>
                             {
                                 (isAccordionOpen) ?
                                     <p className='text-sm'>