+++ packages/wallets/evm/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @layerswap/wallet-evm
 
+## 1.1.2
+
+### Patch Changes
+
+- Fix network switch issue, add rpc health check functionality
+- Updated dependencies
+  - @layerswap/widget@1.1.4
+
 ## 1.1.1
 
 ### Patch Changes