+++ packages/wallets/imtblX/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @layerswap/wallet-imtbl-x
 
+## 1.1.2
+
+### Patch Changes
+
+- Fix wallet connection issue
+- Updated dependencies
+  - @layerswap/widget@1.1.4
+
 ## 1.1.1
 
 ### Patch Changes